+++ packages/mask/src/plugins/WalletService/services/wallet/database/network.ts
@@ -16,8 +16,8 @@ interface Network {
 
 interface AddEthereumChainParameter {
     chainId: string
-    blockExplorerUrls?: string[]
     chainName?: string
+    blockExplorerUrls?: string[]
     iconUrls?: string[]
     nativeCurrency?: {
         name: string
