+++ packages/mask/src/plugins/WalletService/services/wallet/index.ts
@@ -24,6 +24,9 @@ function sanitizeWallet(wallet: Wallet): Wallet {
 // wallet db
 export { hasWallet, updateWallet } from './database/wallet.js'
 
+// network db
+export { getNetwork, getNetworks, addNetwork, updateNetwork, deleteNetwork } from './database/network.js'
+
 // password
 export {
     setPassword,
@@ -35,8 +38,6 @@ export {
     validatePassword,
     clearPassword,
 } from './password.js'
-
-export * from './network.js'
 
 // locker
 export { isLocked, lockWallet, unlockWallet } from './locker.js'
