+++ packages/mask/src/utils/components/Avatar.tsx
@@ -1,5 +1,6 @@
-import { type ProfileInformation as Profile, generateContactAvatarColor } from '@masknet/shared-base'
+import { type ProfileInformation as Profile } from '@masknet/shared-base'
 import { useTheme, Avatar as MuiAvatar, type AvatarProps } from '@mui/material'
+import { generateContactAvatarColor } from '../generateContactAvatarColor.js'
 
 interface Props extends AvatarProps {
     person: Profile
