+++ packages/mask/src/utils/generateContactAvatarColor.ts
