+++ packages/mask/src/utils/index.ts
@@ -13,3 +13,4 @@ export * from './theme/index.js'
 export * from './type-transform/index.js'
 export * from './utils.js'
 export * from './watcher.js'
+export * from './generateContactAvatarColor.js'
