+++ packages/shared-base/src/color/index.ts
@@ -1,5 +1,3 @@
-export * from './avatar.js'
-
 export type Color =
     | `rgb(${number}, ${number}, ${number})`
     | `rgba(${number}, ${number}, ${number}, ${number})`
