+++ packages/shared-base/src/utils/index.ts
@@ -13,6 +13,7 @@ export * from './resolve.js'
 export * from './subscription.js'
 export * from './getDomainSystem.js'
 export * from './parseURL.js'
+export * from './generateContactAvatarColor.js'
 
 export type PartialRequired<T, RequiredKeys extends keyof T> = Omit<T, RequiredKeys> & Pick<Required<T>, RequiredKeys>
