+++ packages/types/src/SocialNetworkUI.ts
@@ -2,7 +2,6 @@ import type { Subscription } from 'use-subscription'
 import type {
     ValueRef,
     EncryptionTargetType,
-    EnhanceableSite,
     NextIDPlatform,
     ObservableWeakMap,
     PersonaIdentifier,
@@ -11,45 +10,14 @@ import type {
     ProfileInformation,
 } from '@masknet/shared-base'
 import type { PaletteMode, Theme } from '@mui/material'
-import type { GrayscaleAlgorithm, SocialNetworkEnum } from '@masknet/encryption'
-import type { IdentityResolved, PostContext, PostContextCreation, PostInfo } from '@masknet/plugin-infra/content-script'
+import type { GrayscaleAlgorithm } from '@masknet/encryption'
+import type { IdentityResolved, PostInfo } from '@masknet/plugin-infra/content-script'
 import type { SerializableTypedMessages } from '@masknet/typed-message'
 import type { RenderFragmentsContextType } from '@masknet/typed-message-react'
 import type { SharedComponentOverwrite } from '@masknet/shared'
 import type { ThemeSettings } from '@masknet/web3-shared-base'
+import type { SocialNetwork } from './ScoialNetwork.js'
 
-export declare namespace SocialNetwork {
-    export interface Utils {
-        /** @returns post URL from PostIdentifier */
-        getPostURL?(post: PostIdentifier): URL | null
-        getUserIdentity?(useId: string): Promise<IdentityResolved | undefined>
-        /** Is this username valid in this network */
-        isValidUsername?(username: string): boolean
-        /** Handle share */
-        share?(text: string): void
-        createPostContext: (creation: PostContextCreation) => PostContext
-    }
-    export interface Shared {
-        utils: Utils
-    }
-    export interface Base {
-        /**
-         * This name is used internally and should be unique.
-         *
-         * !!! THIS SHOULD NOT BE USED TO CONSTRUCT A NEW ProfileIdentifier !!!
-         */
-        networkIdentifier: EnhanceableSite
-        encryptionNetwork: SocialNetworkEnum
-        /**
-         * This field _will_ be overwritten by SocialNetworkUI.permissions
-         */
-        declarativePermissions: SocialNetworkUI.DeclarativePermission
-        /** Should this UI content script activate? */
-        shouldActivate(location: Location | URL): boolean
-        /** This provider is not ready for production, Mask will not use it in production */
-        notReadyForProduction?: boolean
-    }
-}
 export namespace SocialNetworkUI {
     export interface DeferredDefinition extends SocialNetwork.Base {
         /**
