+++ packages/types/src/index.ts
@@ -1,1 +1,2 @@
-export * from './types.js'
+export * from './ScoialNetwork.js'
+export * from './SocialNetworkUI.js'
